# improve performance and memory consumption