--- a/ansicolor.js
+++ b/ansicolor.js
@@ -37,13 +37,6 @@ const colorCodes      = [   'black',      'red',      'green',      'yellow',   
 
 /*  ------------------------------------------------------------------------ */
 
-const clean = obj => {
-                for (const k in obj) { if (!obj[k]) { delete obj[k] } }
-                return (O.keys (obj).length === 0) ? undefined : obj
-            }
-
-/*  ------------------------------------------------------------------------ */
-
 class Color {
 
     constructor (background, name, brightness) {
@@ -57,10 +50,20 @@ class Color {
         return new Color (!this.background, this.name || (this.background ? 'black' : 'white'), this.brightness)
     }
 
-    get clean () {
-        return clean ({ name:   this.name === 'default' ? '' : this.name,
-                        bright: this.brightness === Code.bright,
-                        dim:    this.brightness === Code.dim })
+    get clean() {
+      const name = this.name === "default" ? "" : this.name;
+      const bright = this.brightness === Code.bright;
+      const dim = this.brightness === Code.dim;
+
+      if (!name && !bright && !dim) {
+        return undefined;
+      }
+
+      return {
+        name,
+        bright,
+        dim,
+      };
     }
 
     defaultBrightness (value) {
@@ -88,23 +91,35 @@ class Color {
 
 class Code {
 
-    constructor (n) {
-        if (n !== undefined) { this.value = Number (n) } }
+  constructor(n) {
+    let value = undefined;
+    let type = undefined;
+    let subtype = undefined;
+    let str = "";
+    let isBrightness = false;
 
-    get type () {
-       return types[Math.floor (this.value / 10)] }
+    if (n !== undefined) {
+      value = Number(n);
+      type = types[Math.floor(value / 10)];
+      subtype = subtypes[type][value % 10];
+      str = "\u001b[" + value + "m";
+      isBrightness =
+        value === Code.noBrightness ||
+        value === Code.bright ||
+        value === Code.dim;
+    }
 
-    get subtype () {
-        return subtypes[this.type][this.value % 10] }
+    this.value = value;
+    this.type = type;
+    this.subtype = subtype;
+    this.str = str;
+    this.isBrightness = isBrightness;
+  }
 
-    get str () {
-        return (this.value ? ('\u001b\[' + this.value + 'm') : '') }
-
-    static str (x) {
-        return new Code (x).str }
-
-    get isBrightness () {
-        return (this.value === Code.noBrightness) || (this.value === Code.bright) || (this.value === Code.dim) }
+  static str(x) {
+    if(x === undefined) return "";
+    return "\u001b[" + Number(x) + "m";
+  }
 }
 
 /*  ------------------------------------------------------------------------ */
@@ -198,48 +213,195 @@ const TEXT    = 0,
       BRACKET = 1,
       CODE    = 2
 
-function rawParse (s) {
-    
-    let state = TEXT, buffer = '', text = '', code = '', codes = []
-    let spans = []
+class Span {
+  constructor(code, text) {
+    this.code = code;
+    this.text = text;
 
-    for (let i = 0, n = s.length; i < n; i++) {
+    // Those are added in the actual parse, this is done for performance reasons to have the same hidden class
+    this.css = "";
+    this.color = "";
+    this.bgColor = "";
+    this.bold = undefined;
+    this.inverse = undefined;
+    this.italic = undefined;
+    this.underline = undefined;
+    this.bright = undefined;
+    this.dim = undefined;
+  }
+}
 
-        const c = s[i]
+// getString as function instead of string to allow garbage collection
+function* rawParse(getString) {
+  const stateObject = {
+    state: TEXT,
+    buffer: "",
+    text: "",
+    code: "",
+    codes: [],
+  };
 
-        buffer += c
+  const ONE_MB = 1048576;
 
-        switch (state) {
+  const chunks = splitStringToChunksOfSize(getString(), ONE_MB);
 
-            case TEXT:
-                if (c === '\u001b') { state = BRACKET; buffer = c; }
-                else                { text += c }
-                break
+  while (chunks.length > 0) {
+    const chunk = chunks.shift();
+    yield* processChunk(chunk, stateObject);
+  }
 
-            case BRACKET:
-                if (c === '[') { state = CODE; code = ''; codes = [] }
-                else           { state = TEXT; text += buffer }
-                break
+  if (stateObject.state !== TEXT) stateObject.text += stateObject.buffer;
 
-            case CODE:
+  if (stateObject.text) {
+    yield new Span(new Code(), stateObject.text);
+  }
+}
 
-                if ((c >= '0') && (c <= '9')) { code += c }
-                else if (c === ';')           { codes.push (new Code (code)); code = '' }
-                else if ((c === 'm'))         { code = code || '0'
-                                                codes.push (new Code (code))
-                                                for (const code of codes) { spans.push ({ text, code }); text = '' }
-                                                state = TEXT
-                                              }
-                else                          { state = TEXT; text += buffer }
+function splitStringToChunksOfSize(str, chunkSize) {
+  const chunks = [];
+  const chunksLength = Math.ceil(str.length / chunkSize);
+
+  for (let i = 0, o = 0; i < chunksLength; ++i, o += chunkSize) {
+    chunks.push(str.substring(o, o + chunkSize));
+  }
+
+  return chunks;
+}
+
+function* processChunk(chunk, stateObject) {
+  const chars = chunk;
+  const charsLength = chunk.length;
+
+  for (let i = 0; i < charsLength; i++) {
+    const c = chars[i];
+
+    stateObject.buffer += c;
+
+    switch (stateObject.state) {
+      case TEXT:
+        if (c === "\u001b") {
+          stateObject.state = BRACKET;
+          stateObject.buffer = c;
+        } else {
+          stateObject.text += c;
+        }
+        break;
+
+      case BRACKET:
+        if (c === "[") {
+          stateObject.state = CODE;
+          stateObject.code = "";
+          stateObject.codes = [];
+        } else {
+          stateObject.state = TEXT;
+          stateObject.text += stateObject.buffer;
+        }
+        break;
+
+      case CODE:
+        if (c >= "0" && c <= "9") {
+          stateObject.code += c;
+        } else if (c === ";") {
+          stateObject.codes.push(new Code(stateObject.code));
+          stateObject.code = "";
+        } else if (c === "m") {
+          stateObject.code = stateObject.code || "0";
+          for (const code of stateObject.codes) {
+            yield new Span(code, stateObject.text);
+            stateObject.text = "";
+          }
+
+          yield new Span(new Code(stateObject.code), stateObject.text);
+          stateObject.text = "";
+          stateObject.state = TEXT;
+        } else {
+          stateObject.state = TEXT;
+          stateObject.text += stateObject.buffer;
         }
     }
-
-    if (state !== TEXT) text += buffer
-
-    if (text) spans.push ({ text, code: new Code () })
-
-    return spans
+  }
 }
+
+
+/**
+ * Parse ansi text
+ * @param {Generator<Span, void, *>} rawSpansIterator raw spans iterator
+ * @return {Generator<Span, void, *>}
+ */
+function* parseAnsi(rawSpansIterator) {
+    let color = new Color();
+    let bgColor = new Color(true /* background */);
+    let brightness = undefined;
+    let styles = new Set();
+
+    function reset() {
+        color = new Color();
+        bgColor = new Color(true /* background */);
+        brightness = undefined;
+        styles.clear();
+    }
+
+    reset();
+
+    for (const span of rawSpansIterator) {
+        const c = span.code;
+
+        const inverted = styles.has("inverse");
+        const underline = styles.has("underline")
+            ? "text-decoration: underline;"
+            : "";
+        const italic = styles.has("italic") ? "font-style: italic;" : "";
+        const bold = brightness === Code.bright ? "font-weight: bold;" : "";
+
+        const foreColor = color.defaultBrightness(brightness);
+
+        span.css = bold + italic + underline + foreColor.css(inverted) + bgColor.css(inverted);
+        span.bold = !!bold;
+        span.color = foreColor.clean;
+        span.bgColor = bgColor.clean;
+        span.inverse = inverted;
+        span.italic = !!italic;
+        span.underline = !!underline;
+        span.bright = styles.has("bright");
+        span.dim = styles.has("dim");
+
+        yield span;
+
+        if (c.isBrightness) {
+            brightness = c.value;
+            continue;
+        }
+
+        if (span.code.value === undefined) {
+            continue;
+        }
+
+        if (span.code.value === Code.reset) {
+            reset();
+            continue;
+        }
+
+        switch (span.code.type) {
+            case "color":
+            case "colorLight":
+                color = new Color(false, c.subtype);
+                break;
+
+            case "bgColor":
+            case "bgColorLight":
+                bgColor = new Color(true, c.subtype);
+                break;
+
+            case "style":
+                styles.add(c.subtype);
+                break;
+            case "unstyle":
+                styles.delete(c.subtype);
+                break;
+        }
+    }
+}
+
 
 /*  ------------------------------------------------------------------------ */
 
@@ -252,8 +414,7 @@ class Colors {
      * @param {string} s a string containing ANSI escape codes.
      */
     constructor (s) {
-
-        this.spans = s ? rawParse (s) : []
+        this.spans = s ? Array.from(rawParse(typeof s === 'string' ? () => s : s)) : []
     }
 
     get str () {
@@ -261,68 +422,11 @@ class Colors {
     }
 
     get parsed () {
+        const newColors = new Colors();
 
-        let color, bgColor, brightness, styles
+        newColors.spans = Array.from(parseAnsi(this.spans));
 
-        function reset () {
-
-            color      = new Color (),
-            bgColor    = new Color (true /* background */),
-            brightness = undefined,
-            styles     = new Set ()
-        }
-
-        reset ()
-
-        return O.assign (new Colors (), {
-
-            spans: this.spans.map (span => {
-
-                const c = span.code
-
-                const inverted  = styles.has ('inverse'),
-                      underline = styles.has ('underline')   ? 'text-decoration: underline;' : '',                      
-                      italic    = styles.has ('italic')      ? 'font-style: italic;' : '',
-                      bold      = brightness === Code.bright ? 'font-weight: bold;' : ''
-
-                const foreColor = color.defaultBrightness (brightness)
-
-                const styledSpan = O.assign (
-                                        { css: bold + italic + underline + foreColor.css (inverted) + bgColor.css (inverted) },
-                                        clean ({ bold: !!bold, color: foreColor.clean, bgColor: bgColor.clean }),
-                                        span)
-
-                for (const k of styles) { styledSpan[k] = true }
-
-                if (c.isBrightness) {
-
-                    brightness = c.value
-                
-                } else if (span.code.value !== undefined) {
-
-                    if (span.code.value === Code.reset) {
-                        reset ()
-
-                    } else {
-
-                        switch (span.code.type) {
-
-                            case 'color'        :
-                            case 'colorLight'   : color   = new Color (false, c.subtype); break
-
-                            case 'bgColor'      :
-                            case 'bgColorLight' : bgColor = new Color (true,  c.subtype); break
-
-                            case 'style'  : styles.add    (c.subtype); break
-                            case 'unstyle': styles.delete (c.subtype); break
-                        }
-                    }
-                }
-
-                return styledSpan
-
-            }).filter (s => s.text.length > 0)
-        })
+        return newColors;
     }
 
 /*  Outputs with Chrome DevTools-compatible format     */
@@ -360,6 +464,15 @@ class Colors {
      */
     static parse (s) {
         return new Colors (s).parsed
+    }
+
+    /**
+     *
+     * @param {string | () => string} s string or a function returning a string (for large strings you may want to use a function to avoid memory issues)
+     * @returns {Generator<Span, void, *>} Spans iterator
+     */
+    static parseIterator(s) {
+        return parseAnsi(rawParse(typeof s === "string" ? () => s : s));
     }
 
     /**