--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ansicolor",
-  "version": "1.1.100",
+  "version": "2.0.0",
   "description": "A JavaScript ANSI color/style management. ANSI parsing. ANSI to CSS. Small, clean, no dependencies.",
   "main": "./build/ansicolor.js",
   "types": "./ansicolor.d.ts",