… build/ansicolor.js
… test.js