Improve performance and memory consumption drastically

| Size  | Before memory | Before duration | After memory | After duration |
|-------|---------------|-----------------|--------------|----------------|
| 4MB   | 226MB         | 260ms           | 86 MB        | 167ms          |
| 62MB  | 2.42GB        | 4.4s            | 286 MB       | 1.159s         |
| 150MB | 5.67GB        | 14s             | 565 MB       | 2.75s          |
| 317MB | 11.9GB        | 42s             | 1.09GB       | 5.795s         |


(the benchmark for the after used the parseIterator)

[benchmark](https://github.com/rluvaton/ansi-color-benchmark)



Migrated after improving the performance in my [colorize ansi chrome extension](https://github.com/rluvaton/colorize-ansi-browser-extension) that bundled this library