@xpl 2023-11-07
@rluvaton I am getting troubles with running the bundled .js file (can be reproduced with `npm run test`):

```
ReferenceError: regeneratorRuntime is not defined
    at Object.<anonymous> (/Users/v-gordon/ansicolor/build/ansicolor.js:1:59011)
    at Module._compile (module.js:653:30)
    at Module.replacementCompile (/Users/v-gordon/ansicolor/node_modules/append-transform/index.js:58:13)
    at Module._extensions..js (module.js:664:10)
    ```
    
I am quite sure it is because of generators usage, as babel tries to compile it to ES5 and fails to include some polyfill for emulation of generators or something...

But it's almost 2024 and I am not sure we still need to compile packages to ES5, seems ancient! What do you think if we remove the `build` and will just refer to the source `.js` file without any transpilation? Could it break anything?