@rluvaton 2023-11-07
Sounds great, could break older browsers of course but I think it's fine