@xpl 2023-11-07
> Should I bump the version as well?

@rluvaton Yes, bump plz to 2.0.0 — as we lose the compatibility with older browsers, we don't want ppl to get autoupdated to it.

I'll publish