@xpl 2023-11-07
> You should really use CI/CD for running the test + build + deployed with auto bumping

I originally used Travis CI for that (there was a CD script, you can see it's remnants in the repo), but it had been breaking occasionally and I didn't have the time to maintain it, so I dropped it — as manually publishing wasn't really an issue.

Since we now have Github actions, it would be reasonable to add GH-based CD, as I guess the maintenance burden will be lower than with Travis. I still don't have the time for that, so if you desire to make a PR, it would be awsome.