@xpl 2023-11-07
@rluvaton Running the bundled `.js` now works flawlessly — thanks for that!

But now I noticed that some tests are failing — could you please investigate and fix whatever is causing that?

You could reproduce it by running `npm run test`