@rluvaton 2023-11-07
Most of the tests because of missing properties that now always shown and one weird failure (checking now)