@JonasBa 2023-11-07
@rluvaton I would recommend re-running the original benchmark posted in your PR description so we can see if any of the final changes impacted it. Awesome work 👏🏼